haraikomi-OCR/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
numpy>=1.26.0
Pillow>=10.2.0
yomitoku>=0.7.0
google-api-python-client>=2.165.0
google-auth-oauthlib>=1.2.0
